Refactored the notification
