+++ packages/js/src/components/TimeConstrainedNotification.js
@@ -1,7 +1,6 @@
 import PropTypes from "prop-types";
 import { select } from "@wordpress/data";
 import PersistentDismissableNotification from "../containers/PersistentDismissableNotification";
-import { ReactComponent as DefaultImage } from "../../../../images/succes_marieke_bubble_optm.svg";
 
 /**
  * Checks if there are any warnings.
@@ -20,15 +19,17 @@ const areThereAnyWarnings = () => {
  * @param {JSX.Element} image The image or null if no image.
  * @param {string}      title The title of the notification.
  * @param {string}      promoId The promotion id.
+ * @param {string}      alertKey The unique id for the alert.
  * @param {JSX.node}    children The content of the notification.
  *
  * @returns {JSX.Element} The TimeConstrainedNotification component.
  */
 export const TimeConstrainedNotification = ( {
 	store = "yoast-seo/editor",
-	image: Image = DefaultImage,
+	image: Image = null,
 	title,
 	promoId,
+	alertKey,
 	children,
 	...props
 } ) => {
@@ -37,9 +38,9 @@ export const TimeConstrainedNotification = ( {
 
 	return (
 		! areThereAnyWarnings() && promotionActive && <PersistentDismissableNotification
-			alertKey="time-constrained-notification"
+			alertKey={ alertKey }
 			store={ store }
-			id="time-constrained-notification"
+			id={ alertKey }
 			title={ title }
 			image={ Image }
 			{ ...props }
@@ -54,5 +55,6 @@ TimeConstrainedNotification.propTypes = {
 	image: PropTypes.elementType,
 	title: PropTypes.string.isRequired,
 	promoId: PropTypes.string.isRequired,
+	alertKey: PropTypes.string.isRequired,
 	children: PropTypes.node,
 };
